Disable long_description warnings, enable public method warnings.
The minimum required number of public functions was changed to 2.
This is to prevent the proliferation of classes that could better
be expressed as a function closure. Some classes, such as
enums or utility classes which only override internal methods
will still exist - warnings for those can be disabled using
inline suppressions.