Add tests for util.sh using BATS.
The tests are run by invoking "bats" on a file ending in ".bats".
These files will use the special BATS "load" command to load
tests/polysquare_ci_scripts_helper.bash.

The tests at present test most of the shared functionality
contained in util.sh, including the monitoring of commands and
reporting of failures. They also include the utility functions
to forward arguments.

In util-fetch.bats, a special fixture is used to start a webserver
which will serve a simple bash file. This is used to test the
polysquare_fetch_and_* family of functions.