Added shell-check.bats, with tests shell linting and testing.
These tests ensure that we are able to run the linters on
shell, bash and BATS files (after they have been preprocessed).

It turns out that bash files and bats files were never being
considered for linting due to a quoting issue, so this was
fixed along with the tests.

Finally, the capacity to lint bats files was also added, by
pre-processing them and then passing the temporary results
to bats.