Rewrite ci-scripts in python.
This rewrite follows a similar structure to the bash-scripts that
preceeded these ones. Setup scripts are placed in ciscripts/setup,
check scripts in ciscripts/check.

The internals are written in python, but the scripts themselves
emit bash commands to save state between invocations.

Utility functions are found both in bootstrap.py and util.py

At the moment, the scripts only handle python projects, but this
will change in future as more complex scripts come online.