Ensure that returned python path is absolute.
Previously we were just getting the last component, which was a
relative path to the python installation from a directory
that we were not actually in at the time. Join the installation
so that we get the absolute path.

This fixed removal of files in python installations, which
previously failed silently.