Support both ":" and ";" as a variable delimiter.
On Windows, environment variable lists (for variables like
PATH) are delimited by ";" instead of ":" because ":" is used by
drive letters. As such, we need to transparently convert back
and forth between the two.

Bash still uses ":" as its delimiter, so ";" separated variables
need to be converted to use ":" instead. Additionally, Bash
doesn't support drive-letter like path names and maps them to
/(letter)/path instead. As such, paths should be converted to use
that notation when being passed to Bash, but keep the drive letter
notation everywhere else.