Process shebangs and path extensions in execute.
Windows doesn't internally pre-process shebangs in text files
before executing them, so if a file has no extension and is not
registered with PATHEXT then it won't execute.

We need to open the file and read it shebang, then modify the
command line to run the file with its interpreter.

In addition, anything in PATHEXT can be executed without its
extension by way of shorthand, so this case needs to be handled
in which() as well.