Use a queue to communicate between threads
Previously we used select and anonymous pipes to signal
to the output thread that the subprocess had exited. Windows
does not support non-blocking reads or a select() like interface.

As such, we should just use a queue, which does locking internally
to ensure that messages are safetly passed between threads.