Ensure that bootstrap.__file__ is an absolute path.
During tests, we read bootstrap.__file__ to determine where to
copy it from so that we can have a copy in our temporary container.

If __file__ is a relative as opposed to an absolute path, then
changing into a new directory before reading __file__ will give
an incorrect path.

Patch __file__ on setUp to ensure that the path is absolute.