Handle output from config.guess on Windows.
On Windows, config.guess will return "mingw" as the platform
as opposed to "windows."

If the python binary is a different architecture than the target
system, then its likely that config.guess will pick up the
architecture with the lowest bitness, as python was compiled
against that mingw. In that case, we should just skip the test.