Add separate python installer for Windows and Unix.
On Windows, pyenv isn't supported, so we just install python
directly from its binary installer.

On Unix, we use pyenv.

Also note that on Windows, the install structure is a little
different - PYTHONPATH should be set to Lib/site-packages as
opposed to lib/python-version/site-packages. The expectation
appears to be that each python version is installed in a separate
place. Finally, scripts like pip and easy_install are
placed in Scripts as opposed to bin, so adjust PATH accordingly.