Pick method of printing based on parent shell
Differnent parent shells will have different requirements
in terms of what can and cannot be printed.

In bash, status messages should be printed to the standard error
and escaped with a backslash.

In powershell, status messages should be printed to the standard
output, script code to a file, escaped with a backtick.

This needs to be specified by the user on the command line,
including any relevant file which script should be printed to.