Allow preinstalled ruby and python installations
This saves time from having to install python and ruby
if we have a system installation that we can use directly.

No tests are performed on the systme installation to make sure
that it is the write version or is writable by the user.

This is useful on platforms like Windows where installing the same
version of some software twice isn't possible because that version
was already in the registry under a different PATH