Ensure that _fetch_script runs during the tests.
It turns out that scripts weren't being fetched from our
HTTP server, since it ran in the same directory as the
scripts were located, causing the optimisation to fetch
them locally instead to kick in.

Change into a new directory and fix up the SimpleHTTPRequestHandler
derivative to change back into the directory where the files
should be served from before serving them, to ensure that
they are actually served.