Use OS-specific shell in tests
For Windows, that is powershell, on Linux that is bash.

The reason for this is that we don't want to cause python to
read environment variables directly from the parent shell's
exports and then get values with the wrong delimiter. Bash
requires that environment variables are delimited with ":".

Python inherits the old values by default, but if they've been
changed on the bash side, then bash's delimiters will prevail
and our paths won't be separated by os.pathsep anymore.