Constrain search for existing packages to container paths
In most cases, we already pass an executable path (path=) to
util.where_unavailable, so use this information to constrain
the search for programs to that path only. This ensures that
we always install the requested program in our container even
if it exists in some other path that is in the PATH variable.