Make sure file passed to Popen is an absolute path.
On Windows, PATH is checked in its state as it existed
when the parent process started and not at the time Popen
is called. This diverges from the position on POSIX, where
the latter is true. We rely on the latter behaviour, so
explicitly call which() to ensure that we get the
absolute path to the executable we want to run.