bii: Avoid running "bii" all the time
We only need to use the bii wrapper when files that are required
to build a bii project don't exist - at all other times we can
emulate what the bii executable does when building a cmake project.

This saves about ten seconds of build time where the bii executable
kept on checking for new dependencies on each invocation.