container: Reference-count when we enter language containers
This fixes subtle problems when contaienrs are entered twice
and then deactiated. If container A is entered and then entered
A again, then exited, we are still in A. Previously, the entire
container would have been deactivated when the user would have
expected that it was activated. Reference count to get around
this problem.